requirements.txt
pymongo~=4.8.0
loguru~=0.7.2
python-dotenv~=1.0.1